changed the styles in the buttons
