+++ src/Components/Botones/BotonAgregar/BotonAgregar.css
@@ -1,12 +1,16 @@
 .botonAgregar-neumorphism {
-  padding: 0.625rem 1.25rem;
+  padding: 1rem;
   margin-left: 6.25rem;
   font-size: 1.25rem;
-  font-weight: 700;
+  width: 3rem;
+  height: 3rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   color: white;
   background-color: var(--color-myYellow);
-  border-radius: 1.4rem;
-  box-shadow: 29px 29px 58px #c78b1a, -29px -29px 58px #ffbb23;
+  border-radius: 50%;
+  box-shadow: 5px 5px 38px #c78b1a, -5px -5px 18px #ffbb23;
   cursor: pointer;
   transition: all 300ms;
 }
