+++ src/Components/Botones/BotonAgregar/BotonAgregarSinNeu/BotonAgregarsinNeu.tsx
@@ -1,0 +1,13 @@
+import './BotonAgregarsinNeu.css';
+
+export const BotonAgregarsinNeu = () => {
+  return (
+    <>
+      <button
+        type="button"
+        className="botonAgregar-sin-neumorphism text-4xl">
+        +
+      </button>
+    </>
+  );
+};
